Move property_row.dart to a common directory
