+++ example/lib/common/components/property_row.dart
